Added checks for ns and vm when cleaning up (#67309)

+++ ci-operator/step-registry/ipi/deprovision/vsphere/virt/ipi-deprovision-vsphere-virt-commands.sh
@@ -15,15 +15,21 @@ VM_NAMESPACE="${NAMESPACE}"
 
 VM_NAME="$(oc get infrastructure cluster -o json --kubeconfig=${CLUSTER_KUBECONFIG} | jq -r '.status.infrastructureName')-bm"
 
-# Clean up the VM.  The node can live in the cluster since it will be destroyed shortly.
-echo "$(date -u --rfc-3339=seconds) - Stopping VM"
-virtctl stop "${VM_NAME}" --force --grace-period=0 --kubeconfig="${VIRT_KUBECONFIG}"
+# Check if namspace exists
+if [ "$(oc get ns "${VM_NAMESPACE}" --kubeconfig="${VIRT_KUBECONFIG}" --ignore-not-found)" != "" ]; then
+  # Check and see if BM node exists in
+  if [ "$(oc get vm -n "${VM_NAMESPACE}" "${VM_NAME}" --kubeconfig="${VIRT_KUBECONFIG}" --ignore-not-found)" != "" ]; then
+    # Clean up the VM.  The node can live in the cluster since it will be destroyed shortly.
+    echo "$(date -u --rfc-3339=seconds) - Stopping VM"
+    virtctl stop "${VM_NAME}" --force --grace-period=0 --kubeconfig="${VIRT_KUBECONFIG}"
 
-echo "$(date -u --rfc-3339=seconds) - Deleting VM"
-oc delete -f "${SHARED_DIR}/vm.yaml" --kubeconfig="${VIRT_KUBECONFIG}"
+    echo "$(date -u --rfc-3339=seconds) - Deleting VM"
+    oc delete -f "${SHARED_DIR}/vm.yaml" --kubeconfig="${VIRT_KUBECONFIG}"
+  fi
 
-# Clean up namespace of Openshift Virt server if no addition VMs are present from other test runs
-if [[ "$(oc get vm -n ${VM_NAMESPACE} --kubeconfig="${VIRT_KUBECONFIG}" -o name | wc -l)" -eq 0 ]]; then
-  echo "$(date -u --rfc-3339=seconds) - Removing job namespace"
-  oc delete ns ${VM_NAMESPACE} --kubeconfig="${VIRT_KUBECONFIG}"
+  # Clean up namespace of Openshift Virt server if no addition VMs are present from other test runs
+  if [[ "$(oc get vm -n ${VM_NAMESPACE} --kubeconfig="${VIRT_KUBECONFIG}" -o name | wc -l)" -eq 0 ]]; then
+    echo "$(date -u --rfc-3339=seconds) - Removing job namespace"
+    oc delete ns ${VM_NAMESPACE} --kubeconfig="${VIRT_KUBECONFIG}"
+  fi
 fi
